Reconciler: Add "Show me this card again later" option
When there is an ambiguous match you might want to defer it until after
you've done the rest of the matching.

This change adds an option which can be activated with the '?' key, when
used this puts the card at the back of the stack so that it will be
shown again at the end of reconciliation.