Rename Reconciler as Matcher
I think we have a lot of conceptual confusion here because we're using
'Reconcile' to mean two quite separate things — one the generic act of
reconciling data from two different sources, where the software knows
how to do that; the other the specific "reconciliation interface" where
a _human_ has to provide that mapping. Renaming the 'Reconciler' class
(which is the first of these) to 'Matcher' (so that Reconciler is only
used for the second), may help keep that all cleaner.