Don't abort from Reconciler interface generation
The Reconciler Interface shouldn't care what the workflow is of the
thing that's calling it. Rather than having it check for the
pre-existence of a CSV file, it should happily just return the name of
the HTML file it has generated, and let the rake task decide whether to
abort for the user to run the Reconciliation process, or continue based
on previous data.