Handle corrections files
At the end of the regular build, look for a 'manual/corrections.csv'
that includes manual overrides to source data that we know to be
incorrect. This requires supplying both the 'old' and 'new' value, so
that if the underlying source changes again we don't blindly reapply
what may now be an out of date correction.