Split the Filter file by Membership type
Whilst we're building the Filter files, we should group the Memberships
by what type they are — e.g. the Membership of this or other
legislatures; memberships of the Executive, etc.

This will let us process them differently, when we get around to doing
more (or output them better in Popolo)