Use JSON5 so the filter file can be badly formatted
As the workflow for creating the filter file involves moving JSON hash
elements around, it's time-consuming, error-prone, and generally
annoying to make sure that none of the lists have trailing commas.

So read the file in as JSON5 instead, which is more liberal about this
sort of thing. (Though emit it again at the end as pure JSON, so that
other people can consume it more easily.)  #postel