Ensure gender header is added to merged CSV
When we're constructing the merged CSV file to be given to
CSV-to-Popolo, ensure that we add a gender header when reading data from
GenderBalance, rather than the headers from the Gender-Balance API file.

Without this, we end up with (usually blank) columns for male, female,
etc, and none for 'gender' itself, ensuring that the data is never
recognised.