Update Popolo memberships with Membership Matrix data
If a country has specified a Membership Matrix source then we override
all memberships with the information that it generates.

This does mean that if there are new memberships added in other sources
then they won't appear in the final output because we're overriding
everything, but this is OK because if a country is using the Membership
Matrix then we can treat that as the source of truth.