Convert to Integers from GenderBalance file
We don't turn this on generally, as then anything that looks like a
number gets converted to one, often with odd results (especially if it
starts with a  leading zero). But with the GenderBalance data we know
that everything is either a UUID or an integer, so it's safe to convert
(and will make the later arithmetic much easier).