Reconciler: restrict the clicking area to the header area
We want to be able to do other things in the card that require clicking
(e.g. following links, adding a "show more…" button), so prepare for
this by restricting the "Click to match" area to the header.

(As 99% of usage of the reconciler is currently by me, and I almost
always use the keyboard rather than the mouse, the negative impact of
this should be negligible)