Look up election information from Wikidata
Add a Source and RemoteSource for election information from Wikidata.

This uses a "base" Wikidata ID (e.g. Q22160112: Estonian parliamentary
election) and looks up all Wikidata items that are instances of this.

We don't yet do anything with the resulting information, but this will
at least start gathering it and tracking changes to it over time.