Add 'ignore' option to Patcher
When combining a source, allow the instructions to specify that some
fields should not appear in the output (even if previous sources
haven't provided them).

For example, we generally don't want the names attached to Twitter
handles to flow through (see the UK House of Commons for an example of
this).