Make sure all groups are matched to wikidata results
Previously we were keying the lookup hash by wikidata id, which meant
that the last group which was matched to a given wikidata id would win.

This change ensures that all groups that appear in the
sources/group_wikidata.csv will be matched up to wikidata, even if
multiple different parties are matched to the same id.

The slight downside to this approach is that there is now a lot of
duplication in the `sources/wikidata/groups.json` file, because that
file is keyed by our id for a group. Ideally this would be fixed by just
keying that file by wikidata ids, then matching them up to local ids at
a later point in the process.