Warn of renamed Wikidata items
If we look up the Wikidata for a list of items, and one of those has
been renamed, we will get back the data keyed against the _new_ ID,
rather than the old one. Ideally this would all get resolved
transparently, but for now we can at least cope better with not having
the expected results for an item.