Warn if Morph queries have no ORDER BY
We want the incoming data to be in a consistent order, so that diffs are
more obvious, and we don't get false positives simply from reordering.

So warn any time we have a query that doesn't contain an ORDER BY, so we
can more easily notice and go tidy them up.