Reconciler: Ensure `;` separated images are rendered correctly
Try to split the image field on the `;` and display all images, rather
than just the first one.

There's some unpleasant duplication going on here, but it's not entirely
obvious to me how to pull this out into a separate template. I think it
should be possible to use renderTemplate, but calling that produced an
error, presumably because that function isn't included in the parent
template context? So would be good to clean that up at some stage.