Don't have two different 'terms' hashes
Instead of having a different `term` hash inside and outside the loop
(way too confusing!), build an explicit cache of term end dates (which
is all we use the inner terms data for anyway).

This is also a chance to continue our (very) gradual process of moving
all this to use the EveryPolitician::Popolo library, as v0.5 of that
added a 'terms' method.