add a new UUID Mapping class
Add a new class to represent the `data-ids.cv` files we use to map
incoming source IDs to our UUIDs.

We want to move more of the logic here over time. For now, we just
start with the basic logic we already have of reading the file and
converting to a Hash, and of writing the file back out again from a
Hash.

We could also simplify this a little by using `Pathname`, but for now,
let's just keep the logic as close as possible to the live version.