make Source use the new UUID Mapper
now that we have a UUID Mapping class — use it!

(We have no automated tests for this level of the code, but it runs
cleanly on existing data, and if I remove several lines from the middle
of an existing -ids file, then rebuild, those IDs reappear with new
UUIDs.