Move the check for generating the interface
Rather than the caller working out whether we need to generate the
interface, run the check at the point when we're being asked for the
data. This then makes that call check if it needs to generate an
interface, then check if there's already data, and only then return it,
bailing out in either other case.