Gradually move -ids.csv files
Every Membership source has a parallel `source-ids.csv` file to provide
the mapping between the incoming source IDs, and our matching UUIDs.
Currently those live in the same directory as the source file.

We would like to, instead, move those into their own directory, so that
it's much more obvious that these are a different type of file, and the
source directory only contains files that are actually from the remote
sources.

We can do that by hooking into the UUID Map class to make it aware of
both the old location, and the new one. Then if it can't read from the
new location, it will fall back on the old location — but always
rewrite the file out to the new location. Over time, as we build every
country, the files will gradually all be moved, and we can simplify
this down again.