Make sure test IDs file are in a safe directory
We need to make sure we have write permission to a sibling directory to
where we put our test files. If they just go in /tmp (for example), we
almost certainly won't have write permission in /idmap — so we need to
make make sure they we start in /tmp/manual/ instead.